+++ src/main/java/com/zjxu/educationapp/modules/entity/StuHomework.java
@@ -4,7 +4,6 @@ import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.Builder;
@@ -51,7 +50,7 @@ public class StuHomework {
     /**
      * 作业得分
      */
-    private BigDecimal score;
+    private Double score;
 
     /**
      * 批改时间
